+++ elements/standard/ratedsource.cc
@@ -49,8 +49,8 @@ RatedSource::configure(Vector<String> &conf, ErrorHandler *errh)
 
     String data =
       "Random bullshit in a packet, at least 64 bytes long. Well, now it is.";
-    unsigned rate = 10;
-    unsigned bandwidth = 0;
+    uint64_t rate = 10;
+    uint64_t bandwidth = 0;
     int limit = -1;
     int datasize = -1;
     bool active = true, stop = false;
